+++ chains/maker/erisdb_chains.go
@@ -4,8 +4,8 @@ import (
 	"strings"
 
 	"github.com/eris-ltd/eris-cli/definitions"
+	"github.com/eris-ltd/eris-cli/loaders"
 	"github.com/eris-ltd/eris-cli/log"
-	"github.com/eris-ltd/eris-cli/util"
 )
 
 func MakeErisDBChain(name string, seeds []string, accounts []*definitions.ErisDBAccount, chainImageName string,
@@ -16,7 +16,7 @@ func MakeErisDBChain(name string, seeds []string, accounts []*definitions.ErisDB
 		log.WithFields(log.Fields{
 			"name":    account.Name,
 			"address": account.Address,
-			"tokens":  account.Tokens,
+			"tokens":  account.Amount,
 			"perms":   account.ErisDBPermissions.ErisDBBase.ErisDBPerms,
 		}).Debug("Making an ErisDB Account")
 
@@ -29,17 +29,17 @@ func MakeErisDBChain(name string, seeds []string, accounts []*definitions.ErisDB
 		}
 	}
 	for _, account := range accounts {
-		if err := util.WritePrivVals(genesis.ChainID, account); err != nil {
+		if err := loaders.WritePrivVals(genesis.ChainID, account); err != nil {
 			return err
 		}
-		if err := util.WriteGenesisFile(genesis.ChainID, genesis, account); err != nil {
+		if err := loaders.WriteGenesisFile(genesis.ChainID, genesis, account); err != nil {
 			return err
 		}
 		// TODO: [ben] we can expose seeds to be written into the configuration file
 		// here, but currently not used and we'll overwrite the configuration file
 		// with flag or environment variable in eris-db container
 		theSeeds := strings.Join(seeds, ",") // format for config file (if len>1)
-		if err := util.WriteConfigurationFile(genesis.ChainID, account.Name, theSeeds,
+		if err := loaders.WriteConfigurationFile(genesis.ChainID, account.Name, theSeeds,
 			chainImageName, useDataContainer, exportedPorts, containerEntrypoint); err != nil {
 			return err
 		}
@@ -50,7 +50,7 @@ func MakeErisDBChain(name string, seeds []string, accounts []*definitions.ErisDB
 func MakeErisDBAccount(account *definitions.ErisDBAccount) *definitions.ErisDBAccount {
 	mintAct := &definitions.ErisDBAccount{}
 	mintAct.Address = account.Address
-	mintAct.Amount = account.Tokens
+	mintAct.Amount = account.Amount
 	mintAct.Name = account.Name
 	mintAct.Permissions = account.ErisDBPermissions
 	return mintAct
