+++ util/tar.go
@@ -8,7 +8,7 @@ import (
 	"path"
 	"path/filepath"
 
-	"github.com/fsouza/go-dockerclient/external/github.com/docker/docker/pkg/archive"
+	"github.com/docker/docker/pkg/archive"
 )
 
 // used only for docker cp
